refactor: Rule의 필드와 메서드를 책임에 따라 분리
- Rule 내부에 존재하던 필드와 메서드를 책임에 걸맞는 도메인 내부로
이동하였다.
