+++ src/test/java/blackjack/domain/gambler/DealerTest.java
@@ -1,6 +1,6 @@
 package blackjack.domain.gambler;
 
-import static blackjack.domain.Rule.DEALER_NAME;
+import static blackjack.domain.Round.DEALER_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
